Conflicts:
	Proyecto/cine_unimex.v12.suo

+++ Proyecto/cine_unimex/views/NewPrecios.cs
@@ -9,6 +9,7 @@ using System.Windows.Forms;
 using unimex.lenguajesv.cine.DAO;
 using unimex.lenguajesv.cine.DTO;
 
+
 namespace unimex.lenguajesv.cine.views
 {
     public partial class NewPrecios : Form
@@ -37,7 +38,7 @@ namespace unimex.lenguajesv.cine.views
                     PreciosDTO pdto = new PreciosDTO();
                     pdto.Nombre = txtNombrenew.Text;
                     pdto.Descripcion = txtDescnew.Text;
-                    pdto.Precios = Int32.Parse(txtPrecionew.Text);
+                    pdto.Precios = Single.Parse(txtPrecionew.Text);
                     try
                     {
                         PreciosDAO daoprecio = new PreciosDAO();
